refactor: checkForAir moved to ActionHandler.
